review: fix hull doc comment, split ui_tabs, extract free-cam input helper, add tests
- Fix misleading "triplanar-UV'd" in hull_mesh doc (it uses per-face planar UVs)
- Extract ui_tabs() to ui_tabs.rs, bringing ui_menu.rs from 1110 → 910 lines
- Extract dispatch_free_cam_input() helper from dispatch_shortcuts() to isolate
  the free-cam key-poll and event-strip logic in one named method
- Add free_cam unit tests: strafe perpendicularity, adjust_speed min/max clamping,
  enter() round-trip with a non-zero orbit target
