+++ crates/mogen-studio/src/app/ui_menu.rs
@@ -843,54 +843,8 @@ impl MogenStudioApp {
             }
         }
 
-        // Free-cam fly keys. Same viewport-cursor gate as the gizmo letters so
-        // they win over the editor while the user is looking at the 3D view,
-        // but disengage the moment the cursor moves to the text editor. Held
-        // state drives continuous movement (`key_down`), and the WASD letters'
-        // `Text` events are stripped so a focused editor never accumulates
-        // "wwww"; arrow keys carry no text. Movement itself is applied by the
-        // viewer against the scene scale.
         if free_cam && pointer_in_viewport && !modal_typing {
-            let dt = ctx.input(|i| i.stable_dt);
-            let boost = ctx.input(|i| i.modifiers.shift);
-            let mut mv = egui::Vec2::ZERO;
-            ctx.input_mut(|i| {
-                let down = |k| i.key_down(k);
-                if down(egui::Key::W) || down(egui::Key::ArrowUp) {
-                    mv.x += 1.0;
-                }
-                if down(egui::Key::S) || down(egui::Key::ArrowDown) {
-                    mv.x -= 1.0;
-                }
-                if down(egui::Key::D) || down(egui::Key::ArrowRight) {
-                    mv.y += 1.0;
-                }
-                if down(egui::Key::A) || down(egui::Key::ArrowLeft) {
-                    mv.y -= 1.0;
-                }
-                // Keep the movement letters and arrow nav out of the editor.
-                i.events.retain(|e| {
-                    !matches!(
-                        e,
-                        egui::Event::Text(t)
-                            if matches!(
-                                t.as_str(),
-                                "w" | "W" | "a" | "A" | "s" | "S" | "d" | "D"
-                            )
-                    ) && !matches!(
-                        e,
-                        egui::Event::Key {
-                            key: egui::Key::ArrowUp
-                                | egui::Key::ArrowDown
-                                | egui::Key::ArrowLeft
-                                | egui::Key::ArrowRight,
-                            ..
-                        }
-                    )
-                });
-            });
-            self.viewer
-                .free_cam_fly(glam::vec2(mv.x, mv.y), boost, dt);
+            self.dispatch_free_cam_input(ctx);
         }
 
         let mut hit: Option<ShortcutAction> = None;
@@ -907,204 +861,50 @@ impl MogenStudioApp {
         }
     }
 
-    /// Horizontal browser-style tab strip with one entry per open MOG file.
-    /// Replaces the old "Open" list that lived in the left sidebar.
-    pub(super) fn ui_tabs(&mut self, ui: &mut egui::Ui) {
-        let mut activate: Option<usize> = None;
-        let mut close: Option<usize> = None;
-        let mut close_others_of: Option<usize> = None;
-        let mut close_to_right_of: Option<usize> = None;
-        let mut close_all = false;
-        let mut duplicate: Option<usize> = None;
-        let mut copy_path: Option<String> = None;
-        let mut reveal_path: Option<std::path::PathBuf> = None;
-        let mut new_from_empty_area = false;
-        // Total rect the tab strip gets to draw in. We compare the last item's
-        // right edge against this to detect clicks on the unused tail — that's
-        // where a double-click should mint a fresh MOG file.
-        let strip_rect = ui.available_rect_before_wrap();
-        let mut last_item_right = strip_rect.min.x;
-        let total_tabs = self.files.len();
-        egui::ScrollArea::horizontal()
-            .auto_shrink([false, true])
-            .show(ui, |ui| {
-                ui.horizontal(|ui| {
-                    for (i, f) in self.files.iter().enumerate() {
-                        let selected = i == self.active;
-                        // Prefix with a leading bullet for unsaved buffers so
-                        // dirty state is visible at a glance, not just on the
-                        // trailing edge where a long filename can push it
-                        // out of view.
-                        let mut label = String::new();
-                        if f.dirty {
-                            label.push_str("• ");
-                        }
-                        label.push_str(&f.display_name());
-                        if f.llm_in_flight.is_some() {
-                            label.push_str(" ⟳");
-                        }
-                        let resp = ui.selectable_label(selected, label);
-                        let resp = if f.dirty || f.llm_in_flight.is_some() {
-                            resp.on_hover_text(if f.dirty && f.llm_in_flight.is_some() {
-                                "• unsaved changes · ⟳ AI request in progress"
-                            } else if f.dirty {
-                                "• unsaved changes — Cmd/Ctrl+S to save"
-                            } else {
-                                "⟳ AI request in progress"
-                            })
-                        } else {
-                            resp
-                        };
-                        if resp.clicked() {
-                            activate = Some(i);
-                        }
-                        let has_path = f.path.is_some();
-                        let has_right = i + 1 < total_tabs;
-                        let has_others = total_tabs > 1;
-                        resp.context_menu(|ui| {
-                            if ui.button("Duplicate").clicked() {
-                                duplicate = Some(i);
-                                ui.close_menu();
-                            }
-                            if ui
-                                .add_enabled(has_path, egui::Button::new("Copy path"))
-                                .on_hover_text(if has_path {
-                                    "Copy the absolute path of this MOG file to the clipboard"
-                                } else {
-                                    "Save the MOG file first to give it a path"
-                                })
-                                .clicked()
-                            {
-                                if let Some(p) = &f.path {
-                                    copy_path = Some(p.display().to_string());
-                                }
-                                ui.close_menu();
-                            }
-                            if ui
-                                .add_enabled(has_path, egui::Button::new("Reveal in file system"))
-                                .on_hover_text(if has_path {
-                                    "Open the OS file manager at this MOG file's location"
-                                } else {
-                                    "Save the MOG file first to give it a path"
-                                })
-                                .clicked()
-                            {
-                                if let Some(p) = &f.path {
-                                    reveal_path = Some(p.clone());
-                                }
-                                ui.close_menu();
-                            }
-                            ui.separator();
-                            if ui.button("Close tab").clicked() {
-                                close = Some(i);
-                                ui.close_menu();
-                            }
-                            if ui
-                                .add_enabled(has_others, egui::Button::new("Close others"))
-                                .on_hover_text(
-                                    "Close every other tab. Tabs with \
-                                     unsaved changes are skipped.",
-                                )
-                                .clicked()
-                            {
-                                close_others_of = Some(i);
-                                ui.close_menu();
-                            }
-                            if ui
-                                .add_enabled(has_right, egui::Button::new("Close to the right"))
-                                .on_hover_text(
-                                    "Close every tab to the right of this one. \
-                                     Tabs with unsaved changes are skipped.",
-                                )
-                                .clicked()
-                            {
-                                close_to_right_of = Some(i);
-                                ui.close_menu();
-                            }
-                            if ui
-                                .button("Close all")
-                                .on_hover_text(
-                                    "Close every open tab. Tabs with unsaved \
-                                     changes are skipped.",
-                                )
-                                .clicked()
-                            {
-                                close_all = true;
-                                ui.close_menu();
-                            }
-                        });
-                        // Larger hit-area than `small_button("×")`. Frame is
-                        // off so the X looks the same when idle but registers
-                        // a wider clickable region for trackpad / touch users.
-                        let x_resp = ui
-                            .add(
-                                egui::Button::new(
-                                    egui::RichText::new("×").strong(),
-                                )
-                                .frame(false)
-                                .min_size(egui::vec2(18.0, 18.0)),
-                            )
-                            .on_hover_text("Close tab");
-                        if x_resp.clicked() {
-                            close = Some(i);
-                        }
-                        let sep_resp = ui.separator();
-                        last_item_right = last_item_right
-                            .max(x_resp.rect.right())
-                            .max(sep_resp.rect.right());
-                    }
-                });
-            });
-        // Transparent click-catcher over the empty strip to the right of the
-        // last tab. Double-click opens a new MOG file, mirroring the behaviour
-        // of every major browser / editor tab bar.
-        let empty_left = last_item_right;
-        if empty_left < strip_rect.max.x {
-            let empty_rect = egui::Rect::from_min_max(
-                egui::pos2(empty_left, strip_rect.min.y),
-                egui::pos2(strip_rect.max.x, strip_rect.max.y),
-            );
-            let empty_resp = ui.interact(
-                empty_rect,
-                egui::Id::new("tabs_empty_space"),
-                egui::Sense::click(),
-            );
-            if empty_resp.double_clicked() {
-                new_from_empty_area = true;
+    /// Poll WASD / arrow-key fly input for free-cam mode, consume the key
+    /// events so the text editor never sees them, and advance the camera.
+    /// Only called when free cam is active and the pointer is in the viewport.
+    fn dispatch_free_cam_input(&mut self, ctx: &egui::Context) {
+        let dt = ctx.input(|i| i.stable_dt);
+        let boost = ctx.input(|i| i.modifiers.shift);
+        let mut mv = egui::Vec2::ZERO;
+        ctx.input_mut(|i| {
+            let down = |k| i.key_down(k);
+            if down(egui::Key::W) || down(egui::Key::ArrowUp) {
+                mv.x += 1.0;
             }
-            empty_resp.on_hover_text("Double-click to open a new MOG file");
-        }
-        if new_from_empty_area {
-            self.new_untitled();
-        }
-        if let Some(i) = activate {
-            self.activate(i);
-        }
-        if let Some(i) = duplicate {
-            self.duplicate_file(i);
-        }
-        if let Some(path) = copy_path {
-            ui.output_mut(|o| o.copied_text = path.clone());
-            self.active_mut().status = format!("copied path: {path}");
-        }
-        if let Some(path) = reveal_path {
-            let status = match crate::app::editor_link::reveal_in_os(&path) {
-                Ok(()) => format!("revealed {}", path.display()),
-                Err(e) => format!("reveal failed: {} ({e})", path.display()),
-            };
-            self.active_mut().status = status;
-        }
-        if let Some(i) = close {
-            self.request_close_file(i);
-        }
-        if let Some(i) = close_others_of {
-            self.close_other_tabs(i);
-        }
-        if let Some(i) = close_to_right_of {
-            self.close_tabs_to_right(i);
-        }
-        if close_all {
-            self.close_all_tabs();
-        }
+            if down(egui::Key::S) || down(egui::Key::ArrowDown) {
+                mv.x -= 1.0;
+            }
+            if down(egui::Key::D) || down(egui::Key::ArrowRight) {
+                mv.y += 1.0;
+            }
+            if down(egui::Key::A) || down(egui::Key::ArrowLeft) {
+                mv.y -= 1.0;
+            }
+            // Strip WASD text events and arrow key events so they don't
+            // accumulate in a focused editor or trigger editor navigation.
+            i.events.retain(|e| {
+                !matches!(
+                    e,
+                    egui::Event::Text(t)
+                        if matches!(
+                            t.as_str(),
+                            "w" | "W" | "a" | "A" | "s" | "S" | "d" | "D"
+                        )
+                ) && !matches!(
+                    e,
+                    egui::Event::Key {
+                        key: egui::Key::ArrowUp
+                            | egui::Key::ArrowDown
+                            | egui::Key::ArrowLeft
+                            | egui::Key::ArrowRight,
+                        ..
+                    }
+                )
+            });
+        });
+        self.viewer.free_cam_fly(glam::vec2(mv.x, mv.y), boost, dt);
     }
+
 }
